search subfolders for images
Instead of searching the directory just for subdirectories, the
timelapse model now searches recursively through the subdirectories for
any directory containing jpeg/jpg images.

The model is also forced to recheck for changes every time the main page
is returned to, so that any changes to subdirectories are picked up as
well.