use default names without slashes
The default time formating used forward slashes to separate dates, which
were translated into subdirectories. This change ensures there are no
slashes in the default name based on the time and date (it uses
"yyyy-MM-dd-HH:mm" format).

This also tweaks the validation to notice either forward or backward
slashes, and removes the restrictions on colons.

If the name doesn't validate, the "record" button is hidden.